requirements.txt
flask==2.0.1
cerberus==1.3.4
cachetools==4.2.4
schedule==1.1.0
psutil==5.8.0
google_auth_oauthlib
python-dotenv==0.20.0
quickbelog
quickbeutils
requests~=2.26.0
Werkzeug~=2.0.1